requirements.txt
PyQt5
requests
google-cloud-translate
chardet
lxml